moved project folders to correct location
